+++ src/lib.rs
@@ -110,6 +110,9 @@ pub trait Actionlike: Send + Sync + Copy + Eq + Hash + 'static {
     ///
     /// This is mostly used internally, to enable space-efficient iteration.
     fn get_at(index: usize) -> Option<Self>;
+
+    /// Returns the position in the defining enum of the given action
+    fn index(&self) -> usize;
 }
 
 /// An iterator of [`Actionlike`] actions